got rid of warning

+++ src/scribblearea.cpp
@@ -129,9 +129,9 @@ void ScribbleArea::drawBackground()
     midpixel_x_ = room_width_*pixel_to_meters_/2;
     midpixel_y_ = room_height_*pixel_to_meters_/2;
     qPainter.setPen(QPen(Qt::red, 3.0, Qt::SolidLine, Qt::RoundCap, Qt::RoundJoin));
-    qPainter.drawLine(QPoint{midpixel_x_, midpixel_y_}, QPoint{midpixel_x_+pixel_to_meters_, midpixel_y_});
+    qPainter.drawLine(QPoint{midpixel_x_, midpixel_y_}, QPoint{midpixel_x_+(int)pixel_to_meters_, midpixel_y_});
     qPainter.setPen(QPen(Qt::green, 3.0, Qt::SolidLine, Qt::RoundCap, Qt::RoundJoin));
-    qPainter.drawLine(QPoint{midpixel_x_, midpixel_y_}, QPoint{midpixel_x_, midpixel_y_+pixel_to_meters_});
+    qPainter.drawLine(QPoint{midpixel_x_, midpixel_y_}, QPoint{midpixel_x_, midpixel_y_+(int)pixel_to_meters_});
     update();
 }
